# agent/web/minibot-https-www-qianwen-gen-20260319233410/docs/openapi.yaml
openapi: 3.1.0
info:
  title: MiniBot SSE Chat API
  version: "2.0"
  description: |
    Streamed chat completions powered by Qwen models.
    All endpoints use `text/event-stream` for real-time token delivery.

servers:
  - url: http://127.0.0.1:3000/api
    description: Local development server

paths:
  /v2/chat:
    post:
      summary: Stream chat completions
      description: |
        Initiate a streaming chat completion request.
        Returns Server-Sent Events (SSE) with incremental tokens and final status.
      operationId: streamChatCompletion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [messages]
              properties:
                messages:
                  type: array
                  description: Conversation history, each with `role` ("user"/"assistant") and `content`
                  items:
                    type: object
                    required: [role, content]
                    properties:
                      role:
                        type: string
                        enum: [user, assistant]
                      content:
                        type: string
                model:
                  type: string
                  description: Model identifier (e.g., "qwen-plus", "qwen-max")
                  default: qwen-plus
                file_ids:
                  type: array
                  description: List of file IDs to attach to this request (from /v2/upload)
                  items:
                    type: string
                temperature:
                  type: number
                  minimum: 0
                  maximum: 2
                  default: 0.7
      responses:
        '200':
          description: Server-Sent Events stream
          content:
            text/event-stream:
              schema:
                type: string
                description: |
                  Event stream lines, each prefixed with `event:` and `data:`.
                  Valid events: `message`, `done`, `error`.
                  Example:
                  ```
                  event: message
                  data: {"delta":{"content":"Hello"},"finish_reason":null}

                  event: done
                  data: {"finish_reason":"stop"}
                  ```
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  retry_after:
                    type: integer
                    format: int64
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
      x-cors:
        allowOrigins:
          - http://localhost:3000
          - http://127.0.0.1:3000
        allowCredentials: false
        allowHeaders: ["Content-Type"]

  /v2/upload:
    post:
      summary: Upload a file for use in chat
      description: |
        Accepts multipart/form-data uploads.
        Returns a `file_id` usable in `/v2/chat` requests.
      operationId: uploadFile
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: File uploaded successfully
          content:
            application/json:
              schema:
                type: object
                required: [file_id, filename, size, mime_type]
                properties:
                  file_id:
                    type: string
                    description: Unique ID for referencing this file
                  filename:
                    type: string
                  size:
                    type: integer
                    format: int64
                  mime_type:
                    type: string
        '400':
          description: Invalid or missing file
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string

  /v2/chat/{chat_id}/title:
    patch:
      summary: Update session title
      description: |
        Updates the title of an existing chat session.
        Used by sidebar/history panel.
      operationId: updateChatTitle
      parameters:
        - name: chat_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 128
      responses:
        '200':
          description: Title updated successfully
          content:
            application/json:
              schema:
                type: object
                required: [id, title, updated_at]
                properties:
                  id:
                    type: string
                  title:
                    type: string
                  updated_at:
                    type: string
                    format: date-time
        '400':
          description: Invalid title
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string

  /health:
    get:
      summary: Health check endpoint
      description: |
        Returns basic service health status.
      operationId: healthCheck
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                required: [status, timestamp]
                properties:
                  status:
                    type: string
                  timestamp:
                    type: string
                    format: date-time